Added ability to remove files from upload list, implemented zip extraxction, and began modularizing main.py
